move parser into its own module and grammar docs to crate level
